fix(Climbing): use specific climb action to prevent null error
A recent change in the Interactables introduced a new null Action for
grab types that yield no output. The climbing interactable would use
this null action, but this is really not appropriate as the climb
action yields a result so a new climb action has been added that
is still the same in structure as the null action but it now allows
the code to differentiate that this action is actually doing something.