fix(Facade): move deprecated setting to bottom of references
The PseudoBodyFacade property has been deprecated for some time
and the reference to it has now been moved to the bottom of the
references so it is clearly out of the way and not the main focus
for the Facade settings.

The Climb Target property has been moved to the top, even though it
doesn't need setting, this makes it clear that this is a customisable
field.

All protected fields have also been made public so they can be set
during runtime via code.